[configs: restructure] rename files, simplifying
