--- docs/inspr.md
@@ -1,28 +1,0 @@
-## inspr
-
-A brief description of your application
-
-### Synopsis
-
-A longer description that spans multiple lines and likely contains
-examples and usage of using your application. For example:
-inspr ... .
-
-### Options
-
-```
-  -h, --help   help for inspr
-```
-
-### SEE ALSO
-
-* [inspr deploy](inspr_deploy.md)	 - [Cluster] Deploy Workspace on cluster if no arguments passed assuming that Workspace is current directory.
-* [inspr describe](inspr_describe.md)	 - DescribeApp dApp with it dependencies, channel types and third parties
-* [inspr hub](inspr_hub.md)	 - Commands to operate with Inpsr HUB
-* [inspr init](inspr_init.md)	 - [Workspace] Initialize Inspr workspace or dApp
-* [inspr register](inspr_register.md)	 - Register account to get token for Inspr
-* [inspr remove](inspr_remove.md)	 - [Cluster] Remove deployed dApp from cluster
-* [inspr stats](inspr_stats.md)	 - [Cluster] Return statistics from clusters, by default returns statistics only for running clusters
-* [inspr version](inspr_version.md)	 - Print the version number of Inspr cli
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
