--- docs/inspr_deploy.md
@@ -1,26 +1,0 @@
-## inspr deploy
-
-[Cluster] Deploy Workspace on cluster if no arguments passed assuming that Workspace is current directory.
-
-```
-inspr deploy [appName] [flags]
-```
-
-### Options
-
-```
-  -a, --all                  add all dApps to execution
-  -c, --exclude strings      exclude resources from execution
-  -E, --exclude-reg string   exclude resources by RegExp from execution
-  -h, --help                 help for deploy
-  -i, --include strings      include resources into execution
-  -I, --include-reg string   include resources by RegExp into execution
-  -t, --tag string           Add tag to current execution.
-  -w, --workspace string     Path to Inpsr Workspace, by default looks into current working directory.
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
