--- docs/inspr_hub.md
@@ -1,18 +1,0 @@
-## inspr hub
-
-Commands to operate with Inpsr HUB
-
-### Options
-
-```
-  -h, --help   help for hub
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-* [inspr hub describe](inspr_hub_describe.md)	 - DescribeApp dApp with it dependencies, channel types and third parties
-* [inspr hub find](inspr_hub_find.md)	 - Find dApp in Inpsr Hub
-* [inspr hub get](inspr_hub_get.md)	 - Get dApp from Inspr Hub
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
