--- docs/inspr_hub_describe.md
@@ -1,19 +1,0 @@
-## inspr hub describe
-
-DescribeApp dApp with it dependencies, channel types and third parties
-
-```
-inspr hub describe [appName] [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for describe
-```
-
-### SEE ALSO
-
-* [inspr hub](inspr_hub.md)	 - Commands to operate with Inpsr HUB
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
