--- docs/inspr_hub_find.md
@@ -1,19 +1,0 @@
-## inspr hub find
-
-Find dApp in Inpsr Hub
-
-```
-inspr hub find [dApp name] [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for find
-```
-
-### SEE ALSO
-
-* [inspr hub](inspr_hub.md)	 - Commands to operate with Inpsr HUB
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
