--- docs/inspr_hub_get.md
@@ -1,19 +1,0 @@
-## inspr hub get
-
-Get dApp from Inspr Hub
-
-```
-inspr hub get [dAppName...] [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for get
-```
-
-### SEE ALSO
-
-* [inspr hub](inspr_hub.md)	 - Commands to operate with Inpsr HUB
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
