--- docs/inspr_init.md
@@ -1,20 +1,0 @@
-## inspr init
-
-[Workspace] Initialize Inspr workspace or dApp
-
-```
-inspr init [workspace?] [flags]
-```
-
-### Options
-
-```
-  -a, --app string   Init new DApp (should have also -w where to create)
-  -h, --help         help for init
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
