--- docs/inspr_register.md
@@ -1,23 +1,0 @@
-## inspr register
-
-Register account to get token for Inspr
-
-### Synopsis
-
-Prompt: (acc, pass) -> server -> token
-
-```
-inspr register [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for register
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
