--- docs/inspr_remove.md
@@ -1,26 +1,0 @@
-## inspr remove
-
-[Cluster] Remove deployed dApp from cluster
-
-### Synopsis
-
-By default command tries to remove in 'safe' mode, awaiting to close all connections
-
-```
-inspr remove [id...] [flags]
-```
-
-### Options
-
-```
-  -f, --force              Remove dApp without graceful shutdown of service.
-  -h, --help               help for remove
-  -t, --tag string         Add tag to current execution.
-  -w, --workspace string   Path to Inpsr Workspace, by default looks into current working directory.
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
