--- docs/inspr_stats.md
@@ -1,20 +1,0 @@
-## inspr stats
-
-[Cluster] Return statistics from clusters, by default returns statistics only for running clusters
-
-```
-inspr stats [flags]
-```
-
-### Options
-
-```
-  -f, --full   Return statistics from all clusters
-  -h, --help   help for stats
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
