--- docs/inspr_version.md
@@ -1,19 +1,0 @@
-## inspr version
-
-Print the version number of Inspr cli
-
-```
-inspr version [flags]
-```
-
-### Options
-
-```
-  -h, --help   help for version
-```
-
-### SEE ALSO
-
-* [inspr](inspr.md)	 - A brief description of your application
-
-###### Auto generated by spf13/cobra on 18-Jan-2021
